everything updated except bin
