+++ src/index.html
@@ -22,7 +22,6 @@
         </div>
         <button id="todo-footer-button" type="button" class="link-button js-clear-completed">Clear all completed</button>
     </div>
-    
-    
+    <script defer="defer" src="./index.js"></script>
 </body>
 </html>
