--- src/modules/edit.js
@@ -1,9 +1,0 @@
-function todoEdit(todoArray, indx, newTodo) {
-  todoArray.forEach((todo) => {
-    if (todo.index === Number(indx)) {
-      todo.description = newTodo;
-    }
-  });
-}
-
-export default todoEdit;
