+++ src/modules/renderOnLoad.js
@@ -1,0 +1,9 @@
+import todoAdd from './add';
+
+function renderOnLoad(todoArr) {
+  todoArr.forEach((element, indx) => {
+    todoAdd(element, todoArr);
+  });
+}
+
+export default renderOnLoad;
